+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/red_team_agent/__init__.py
@@ -1,0 +1,11 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
+from enum import Enum
+from typing import List
+from .red_team_agent import RedTeamAgent
+from .attack_strategy import AttackStrategy
+from .attack_objective_generator import AttackObjectiveGenerator, RiskCategory
+
+__all__ = ["RedTeamAgent", "AttackStrategy", "RiskCategory", "AttackObjectiveGenerator"]
