+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/red_team_agent/utils/mock_attack_objective.py
@@ -1,7 +1,6 @@
 # ---------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
-
 """Mock implementation of AttackObjective for testing."""
 
 from typing import Any, Dict, List, Optional
